requirements.txt
speedtest-cli
mysql-connector
pyyaml